+++ src/reactor.rs
@@ -1,0 +1,4 @@
+use reactor_trait::{Reactor, TimeReactor};
+
+pub trait FullReactor: Reactor + TimeReactor {}
+impl<R: Reactor + TimeReactor> FullReactor for R {}